# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Milo API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/auth/native/nonce": {
      "post": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Native Nonce",
        "operationId": "native_nonce_auth_native_nonce_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeChallengeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/native/login": {
      "post": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Native Login",
        "operationId": "native_login_auth_native_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeLoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/native/refresh": {
      "post": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Native Refresh",
        "operationId": "native_refresh_auth_native_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeRefreshInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Revoke Session",
        "operationId": "revoke_session_auth_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/native/nonce": {
      "post": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Native Nonce",
        "operationId": "native_nonce_v1_auth_native_nonce_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeChallengeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/native/login": {
      "post": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Native Login",
        "operationId": "native_login_v1_auth_native_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeLoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/native/refresh": {
      "post": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Native Refresh",
        "operationId": "native_refresh_v1_auth_native_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeRefreshInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "device authentication",
          "sessions"
        ],
        "summary": "Revoke Session",
        "operationId": "revoke_session_v1_auth_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NativeChallengeInput": {
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android"
            ],
            "title": "Platform"
          },
          "device_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Device Name",
            "default": "My device"
          },
          "code_challenge": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "pattern": "^[A-Za-z0-9_-]{43}$",
            "title": "Code Challenge"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "platform",
          "code_challenge"
        ],
        "title": "NativeChallengeInput"
      },
...
      "NativeLoginInput": {
        "properties": {
          "credential": {
            "type": "string",
            "maxLength": 16384,
            "minLength": 1,
            "title": "Credential"
          },
          "nonce": {
            "type": "string",
            "maxLength": 256,
            "minLength": 32,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Nonce"
          },
          "code_verifier": {
            "type": "string",
            "maxLength": 128,
            "minLength": 43,
            "pattern": "^[A-Za-z0-9._~-]+$",
            "title": "Code Verifier"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "credential",
          "nonce",
          "code_verifier"
        ],
        "title": "NativeLoginInput"
      },
...
      "NativeRefreshInput": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 256,
            "minLength": 35,
            "pattern": "^nr_[A-Za-z0-9_-]+$",
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "NativeRefreshInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}